# backend/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the documentation of ReMeLife referral api",
    "version": "1.0.0",
    "title": "ReMeLife Api Docs"
  },
  "host": "https://ws-remelife.s6.staging-host.com/",
  "tags": [
    {
      "name": "Referral",
      "description": "Referral code management"
    },
    {
      "name": "Income",
      "description": "Referral income management"
    },
    {
      "name": "Settings",
      "description": "Referral settings"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/getReferralCode/{userId}": {
      "get": {
        "tags": [
          "Referral"
        ],
        "summary": "Get referral code by user id",
        "description": "Returns referral code ",
        "operationId": "getReferralCode",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to return referral code",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/setIncomeSettings": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "set income for all levels",
        "description": "Returns income settings",
        "operationId": "setIncomeSettings",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "main_amount",
            "in": "body",
            "description": "level 0 amount. for newly added user",
            "required": true,
            "type": "integer",
            "example": 100
          },
          {
            "name": "level_1_per",
            "in": "body",
            "description": "level 1 percentage. for user who refer level 0 user directly",
            "required": true,
            "type": "integer",
            "example": 20
          },
          {
            "name": "level_2_per",
            "in": "body",
            "description": "level 2 percentage. for user who refer level 1 user",
            "required": true,
            "type": "integer",
            "example": 10
          },
          {
            "name": "level_3_per",
            "in": "body",
            "description": "level 3 percentage. for user who refer level 2 user",
            "required": true,
            "type": "integer",
            "example": 5
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid body supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/addWithReferral": {
      "post": {
        "tags": [
          "Referral"
        ],
        "summary": "Add user with referral code",
        "description": "Returns added user ",
        "operationId": "addWithReferral",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "body",
            "description": "newly added user id",
            "required": true,
            "type": "string"
          },
          {
            "name": "referred_by",
            "in": "body",
            "description": "referred by refrral code",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid body supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getMyReferrals/{userId}": {
      "get": {
        "tags": [
          "Referral"
        ],
        "summary": "Get refrred users list by user id",
        "description": "Returns refrred users list  ",
        "operationId": "getMyReferrals",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to get referred users list",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getMyReferralsCount/{userId}": {
      "get": {
        "tags": [
          "Referral"
        ],
        "summary": "Get refrred users count with level by user id",
        "description": "Returns refrred users count with level",
        "operationId": "getMyReferralsCount",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to get referred user count",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getMyBalance/{userId}": {
      "get": {
        "tags": [
          "Income"
        ],
        "summary": "Get balance by user id",
        "description": "Returns balance ",
        "operationId": "getMyBalance",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to return balance",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/withdrawal/{userId}": {
      "post": {
        "tags": [
          "Income"
        ],
        "summary": "add withdrawal amount from user balance",
        "description": "Returns balance",
        "operationId": "withdrawal",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to withdrawal amount",
            "required": true,
            "type": "string"
          },
          {
            "name": "withdrawalAmout",
            "in": "body",
            "description": "withdrawal amout",
            "required": true,
            "type": "integer",
            "example": 100
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid body supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getCreditHistory/{userId}": {
      "get": {
        "tags": [
          "Income"
        ],
        "summary": "Get credit history by user id",
        "description": "Returns credit history ",
        "operationId": "getCreditHistory",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to return credit history",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/getWithdrawalHistory/{userId}": {
      "get": {
        "tags": [
          "Income"
        ],
        "summary": "Get withdrawal history by user id",
        "description": "Returns withdrawal history ",
        "operationId": "getWithdrawalHistory",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to return withdrawal history",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  }
}